+++ test/radar/config_test.rb
@@ -9,7 +9,9 @@ class ConfigTest < Test::Unit::TestCase
 
     context "reporters" do
       setup do
-        @reporter_klass = Class.new
+        @reporter_klass = Class.new do
+          def report(event); end
+        end
       end
 
       teardown do
@@ -23,13 +25,11 @@ class ConfigTest < Test::Unit::TestCase
       should "be able to add reporters" do
         @instance.reporters.use @reporter_klass
         assert !@instance.reporters.empty?
-        assert @instance.reporters.values.first.is_a?(@reporter_klass)
       end
 
       should "be able to add reporters via built-in symbols" do
         @instance.reporters.use :file
         assert !@instance.reporters.empty?
-        assert @instance.reporters.values.first.is_a?(Radar::Reporter::FileReporter)
       end
 
       should "yield the reporter instance if a block is given" do